Add a timingSince command statsd client
So we can send stats metrics, provided a starting point for the timing.
So the users of the Statd client do not need to calculated the duration
passed from a given point in time, and just can pass that point in time
to the client and the duration is calculated automatically.